refactor vue-spa-dev
