+++ identity/auth-vue/pages/login/LoginPage.test.ts
@@ -1,5 +1,8 @@
 import { describe, it, expect } from "vitest";
-import { stubGlobals, mountWithPlugins } from "@saflib/vue-spa-dev/components";
+import {
+  stubGlobals,
+  mountWithPlugins,
+} from "../../../../vue-spa/testing/components.ts";
 import { getElementByString } from "@saflib/vue-spa/test-utils";
 import { type VueWrapper } from "@vue/test-utils";
 import LoginPage from "../../pages/login/LoginPage.vue";
