--- vue-spa-dev/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "@saflib/vue-spa-dev",
-  "description": "Library for testing Vue SPAs within the SAF framework",
-  "type": "module",
-  "dependencies": {
-    "@saflib/vitest": "*",
-    "@saflib/vue-spa": "*",
-    "@vue/test-utils": "^2.0.0",
-    "jsdom": "^26.0.0",
-    "sass-embedded": "^1.0.0",
-    "vite-plugin-vuetify": "^2.0.0",
-    "msw": "^2.0.0",
-    "vue-tsc": "^3.0.0"
-  }
-}
