+++ vue-spa/testing/vitest-config.d.ts
