+++ vue-spa/workflows/query-template.test.ts
@@ -1,7 +1,7 @@
 // @ts-nocheck // TODO: remove this
 import { describe, it, expect } from "vitest";
 import { http, HttpResponse } from "msw";
-import { setupMockServer, withVueQuery } from "@saflib/vue-spa-dev/requests.js";
+import { setupMockServer, withVueQuery } from "@saflib/vue-spa/testing";
 import { useQuery } from "@tanstack/vue-query";
 import { useQueryTemplate } from "./query-template.js";
 import type {
